refactor bmi adapters into their own library
