+++ src/core/hydrolocation/CMakeLists.txt
@@ -9,6 +9,8 @@ target_include_directories(core_hydrolocation PUBLIC
         ${PROJECT_SOURCE_DIR}/include/core/catchment
         ${PROJECT_SOURCE_DIR}/include/forcing
         ${PROJECT_SOURCE_DIR}/include/utilities
+        ${PROJECT_SOURCE_DIR}/include/realizations/catchment
+        ${PROJECT_SOURCE_DIR}/include/geojson
         )
 
 target_link_libraries(core_hydrolocation PUBLIC
